refactor(jazz-tools): trim db.ts and extract cross-tab modules
db.ts had grown to 3324 lines with several distinct concerns inlined.
This drops dead helpers and pulls the cross-tab BroadcastChannel
protocol and OPFS storage-reset coordinator into their own modules.

- Remove backendScopedAuthState (single caller; inlined).
- Remove logAuthModeInDev (dev-only console.info on createDb).
- Remove isLeaderDebugEnabled + logLeaderDebug + 16 callsites
  (debug logging gated by localStorage["jazz:leader-debug"], not
  surfaced anywhere outside the file).
- Collapse identical transformInsertInput/transformUpdateInput
  wrappers into transformInputColumns.
- Extract tab-sync-protocol.ts: 7 message-type interfaces,
  TabSyncMessage union, BroadcastChannelLike,
  resolveBroadcastChannelCtor, isTabSyncMessage.
- Extract storage-reset-coordinator.ts: StorageResetCoordinator
  class owning the cross-tab OPFS-wipe state machine. Db wires it
  via a small StorageResetHost interface; Db.deleteClientStorage
  delegates to coordinator.requestReset(); handleSyncChannelMessage
  forwards reset cases.

db.ts: 3324 -> 2663 lines (-661, -20%).
